Divided Layer into two main categories
Layers are divided into layers that can block updates to lower layers
and layers that can't. This is for when it isn't necessary to use a
blocking layer so one doesn't have to create a custom layer interface to
get around that. The only difference is the update() method has a return
type of boolean for BlockingLayer and void for Layer. Returning true
indicates to the calling method to end the update cycle.